+++ billing/tests/test_lifecycle_webhook.py
@@ -16,6 +16,7 @@ def customer(upcoming_period_end):
     """Customer that is coming up for renewal"""
     user = factories.UserFactory(
         paying=True,
+        customer__customer_id="cus",
         customer__subscription_id="sub",
         customer__current_period_end=upcoming_period_end,
     )
@@ -41,20 +42,25 @@ def test_bad_json(client):
     assert models.StripeEvent.objects.count() == 0
 
 
-def test_unrecognized_type(client):
+def test_unrecognized_type(client, customer):
     """Unrecognized event type"""
     url = reverse("billing:stripe_webhook")
     payload = {
         "id": "evt_test",
         "object": "event",
         "type": "bad.type",
-        "data": {"object": None},
+        "data": {
+            "object": {
+                "customer": customer.customer_id,
+                "subscription": customer.subscription_id,
+            }
+        },
     }
     response = client.post(url, payload, content_type="application/json")
     assert 201 == response.status_code
-    assert (
-        models.StripeEvent.Status.IGNORED == models.StripeEvent.objects.first().status
-    )
+    assert models.StripeEvent.Type.UNKNOWN == models.StripeEvent.objects.first().type
+    assert models.StripeEvent.Status.ERROR == models.StripeEvent.objects.first().status
+    assert "Unrecognized payload_type" in models.StripeEvent.objects.first().note
 
 
 def test_renewed(client, customer):
@@ -72,8 +78,16 @@ def test_renewed(client, customer):
                 # See https://stackoverflow.com/questions/22601521/stripe-webhook-events-renewal-of-subscription
                 # for why we need the billing_reason.
                 "billing_reason": "subscription_cycle",
+                "customer": "cus",
                 "subscription": "sub",
-                "lines": {"data": [{"period": {"end": mock_period_end.timestamp()}}]},
+                "lines": {
+                    "data": [
+                        {
+                            "plan": {"id": "price"},
+                            "period": {"end": mock_period_end.timestamp()},
+                        }
+                    ]
+                },
             }
         },
     }
@@ -98,7 +112,19 @@ def test_payment_failure(client, customer, upcoming_period_end):
         "object": "event",
         "type": "invoice.payment_failed",
         "data": {
-            "object": {"subscription": "sub", "billing_reason": "subscription_cycle"}
+            "object": {
+                "customer": "cus",
+                "subscription": "sub",
+                "billing_reason": "subscription_cycle",
+                "lines": {
+                    "data": [
+                        {
+                            "plan": {"id": "price"},
+                            "period": {"end": upcoming_period_end},
+                        }
+                    ]
+                },
+            }
         },
     }
 
@@ -124,7 +150,12 @@ def test_payment_failure_permanent(client, customer, upcoming_period_end):
         "object": "event",
         "type": "customer.subscription.deleted",
         "data": {
-            "object": {"id": "sub", "status": "canceled", "cancel_at_period_end": False}
+            "object": {
+                "id": "sub",
+                "customer": "cus",
+                "status": "canceled",
+                "cancel_at_period_end": False,
+            }
         },
     }
     response = client.post(url, payload, content_type="application/json")
@@ -173,10 +204,132 @@ def test_link_event_to_user(client, customer):
         "object": "event",
         "type": "customer.subscription.deleted",
         "data": {
-            "object": {"id": "sub", "status": "canceled", "cancel_at_period_end": False}
+            "object": {
+                "id": "sub",
+                "customer": "cus",
+                "status": "canceled",
+                "cancel_at_period_end": False,
+            }
         },
     }
     response = client.post(url, payload, content_type="application/json")
     assert response.status_code == 201
     event = models.StripeEvent.objects.first()
     assert event.user == customer.user
+
+
+def test_non_primary_event_subscription(client, customer):
+    """Non-primary events should get the correct type and user attached but be ignored."""
+    url = reverse("billing:stripe_webhook")
+    payload = {
+        "id": "evt_test",
+        "object": "event",
+        "type": "customer.subscription.updated",
+        "data": {
+            "object": {
+                "id": "sub",
+                "customer": "cus",
+                "cancel_at_period_end": False,
+                "status": "active",
+            },
+            "previous_attributes": {"status": "incomplete"},
+        },
+    }
+    response = client.post(url, payload, content_type="application/json")
+    assert response.status_code == 201
+    event = models.StripeEvent.objects.first()
+    assert event.type == models.StripeEvent.Type.NEW_SUB
+    assert event.primary is False
+    assert event.status == models.StripeEvent.Status.IGNORED
+    assert event.user == customer.user
+
+
+def test_user_not_found(client, mock_stripe_customer):
+    """If a user can't be found, error."""
+    url = reverse("billing:stripe_webhook")
+    mock_stripe_customer.retrieve.return_value.email = "notfound@example.com"
+
+    payload = {
+        "id": "evt_test",
+        "object": "event",
+        "type": "customer.subscription.updated",
+        "data": {
+            "object": {
+                "id": "sub",
+                "customer": "cus",
+                "cancel_at_period_end": False,
+                "status": "active",
+            },
+            "previous_attributes": {"status": "incomplete"},
+        },
+    }
+
+    response = client.post(url, payload, content_type="application/json")
+    assert response.status_code == 201
+    event = models.StripeEvent.objects.first()
+    assert event.type == models.StripeEvent.Type.NEW_SUB
+    assert event.primary is False
+    assert event.status == models.StripeEvent.Status.ERROR
+    assert event.user == None
+    assert "Customer.DoesNotExist" in event.note
+
+
+def test_payment_update_active(client, customer):
+    """An update to a Subscription's payment method does not do anything if the Subscription is
+    active."""
+    url = reverse("billing:stripe_webhook")
+
+    payload = {
+        "id": "evt_test",
+        "object": "event",
+        "type": "customer.subscription.updated",
+        "data": {
+            "object": {
+                "id": "sub",
+                "customer": "cus",
+                "cancel_at_period_end": False,
+                "status": "active",
+            },
+            "previous_attributes": {"default_payment_method": "pm_something"},
+        },
+    }
+
+    response = client.post(url, payload, content_type="application/json")
+    assert response.status_code == 201
+    event = models.StripeEvent.objects.first()
+    assert event.type == models.StripeEvent.Type.UPDATE_PAYMENT_METHOD
+    assert event.status == models.StripeEvent.Status.PROCESSED
+    assert event.user.customer == customer
+
+
+@pytest.mark.parametrize("status", ["incomplete", "past_due"])
+def test_payment_update_and_retry(client, customer, status, mock_stripe_invoice):
+    """An update to a Subscription's payment method when not active automatically retries the last open invoice."""
+    mock_stripe_invoice.list.return_value = {
+        "data": [{"status": "open", "id": "inv_123"}]
+    }
+    url = reverse("billing:stripe_webhook")
+
+    payload = {
+        "id": "evt_test",
+        "object": "event",
+        "type": "customer.subscription.updated",
+        "data": {
+            "object": {
+                "id": "sub",
+                "customer": "cus",
+                "cancel_at_period_end": False,
+                "status": status,
+            },
+            "previous_attributes": {"default_payment_method": "pm_something"},
+        },
+    }
+
+    response = client.post(url, payload, content_type="application/json")
+    assert response.status_code == 201
+    event = models.StripeEvent.objects.first()
+    assert event.type == models.StripeEvent.Type.FIX_PAYMENT_METHOD
+    assert event.status == models.StripeEvent.Status.PROCESSED
+    assert event.user.customer == customer
+    assert mock_stripe_invoice.list.call_count == 1
+    assert mock_stripe_invoice.pay.call_count == 1